requirements.txt
azure-ai-vision==0.15.1b1
python-dotenv==1.0.1
